async location from db fetching

+++ app/src/main/java/cloud/ptl/boardgamecollector/db/entity/Location.java
@@ -6,7 +6,7 @@ import androidx.room.PrimaryKey;
 @Entity
 public class Location {
     @PrimaryKey(autoGenerate = true)
-    public int locationId;
+    public long locationId;
     public int gameLocatedId;
 
     public String name;
